requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.0.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2024.12.14
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cache2==0.2.1
flower==2.0.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==0.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
pyasn1==0.6.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
redis==5.2.1
rfc3986==1.5.0
rsa==4.9
sentry-sdk==2.20.0
setuptools==75.7.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
WTForms==3.1.2
yarl==1.18.3
